+++ tests/TestCase.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace PodPoint\SnsBroadcaster\Tests;
+namespace PodPoint\AwsPubSub\Tests;
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
 use Orchestra\Testbench\TestCase as Orchestra;
-use PodPoint\SnsBroadcaster\SnsBroadcasterServiceProvider;
+use PodPoint\AwsPubSub\AwsPubSubServiceProvider;
+use PodPoint\AwsPubSub\EventServiceProvider;
 
 abstract class TestCase extends Orchestra
 {
@@ -22,7 +23,8 @@ abstract class TestCase extends Orchestra
     protected function getPackageProviders($app)
     {
         return [
-            SnsBroadcasterServiceProvider::class,
+            AwsPubSubServiceProvider::class,
+            EventServiceProvider::class,
         ];
     }
 
@@ -34,16 +36,29 @@ abstract class TestCase extends Orchestra
      */
     protected function getEnvironmentSetUp($app)
     {
+        /** PUB */
         $app['config']->set('broadcasting.default', 'sns');
         $app['config']->set('broadcasting.connections.sns', [
             'driver' => 'sns',
-            'region' => 'eu-west-1',
-            'key' => 'foo-bar',
-            'secret' => 'foo-baz',
+            'key' => 'dummy-key',
+            'secret' => 'dummy-secret',
             'arn-prefix' => 'aws:arn:12345:',
-            'arn-suffix' => '-local',
+            'arn-suffix' => '',
+            'region' => 'eu-west-1',
         ]);
 
+        /** SUB */
+        $app['config']->set('queue.connections.pub-sub', [
+            'driver' => 'sqs-sns',
+            'key' => 'dummy-key',
+            'secret' => 'dummy-secret',
+            'prefix' => 'https://sqs.eu-west-1.amazonaws.com/13245',
+            'queue' => 'default',
+            'suffix' => '',
+            'region' => 'eu-west-1',
+        ]);
+
+        /** DATABASE */
         $app['config']->set('database.default', 'testbench');
         $app['config']->set('database.connections.testbench', [
             'driver'   => 'sqlite',